Deduplicate FlightAware data-service helpers
